RAG-based Customer Support Assistant/requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.2.0

# LangGraph for workflow orchestration
langgraph>=0.2.0

# Vector Store
chromadb>=0.5.0

# PDF Processing
pypdf>=4.0.0

# Web UI
streamlit>=1.38.0

# Environment Management
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
